+++ lib/src/gherkin/models/table.dart
@@ -5,4 +5,12 @@ class Table {
   final TableRow header;
 
   Table(this.rows, this.header);
+
+  void setStepParameter(String parameterName, String value) {
+    rows.forEach((row) {
+      row.setStepParameter(parameterName, value);
+    });
+  }
+
+  Table clone() => Table(rows.map((r) => r.clone()).toList(), header.clone());
 }