+++ lib/src/gherkin/parameters/float_parameter.dart
@@ -2,7 +2,7 @@ import './custom_parameter.dart';
 
 class FloatParameterBase extends CustomParameter<num> {
   FloatParameterBase(String name)
-      : super(name, RegExp("([0-9]+.?[0-9]?)"), (String input) {
+      : super(name, RegExp(r"(-?[0-9]+.?[0-9]*)"), (String input) {
           final n = num.parse(input);
           return n;
         });