+++ lib/src/gherkin/runnables/runnable.dart
@@ -1,9 +1,13 @@
 import './debug_information.dart';
 
 abstract class Runnable {
-  final RunnableDebugInformation debug;
-
-  Runnable(this.debug);
-
+  RunnableDebugInformation _debug;
+  RunnableDebugInformation get debug => _debug;
   String get name;
+
+  Runnable(this._debug);
+
+  void updateDebugInformation(RunnableDebugInformation debugInformation) {
+    _debug = debugInformation;
+  }
 }