+++ lib/src/gherkin/syntax/table_line_syntax.dart
@@ -18,6 +18,7 @@ class TableLineSyntax extends RegExMatchedGherkinSyntax {
     if (syntax is TableLineSyntax || syntax is CommentSyntax) {
       return false;
     }
+
     return true;
   }